Removed single-implementation-with-interface pattern from microservices
