+++ PhotoAppApiAlbums/src/main/java/me/learning/microservices/photoapp/api/albums/service/AlbumsService.java
@@ -1,18 +1,55 @@
 package me.learning.microservices.photoapp.api.albums.service;
 
+import lombok.RequiredArgsConstructor;
+import me.learning.microservices.photoapp.api.albums.data.album.Album;
+import me.learning.microservices.photoapp.api.albums.data.album.AlbumsRepository;
+import me.learning.microservices.photoapp.api.albums.mapper.AlbumMapper;
 import me.learning.microservices.photoapp.api.albums.service.exception.AlbumNotFoundException;
 import me.learning.microservices.photoapp.api.albums.service.exception.AlbumServiceException;
 import me.learning.microservices.photoapp.api.albums.shared.AlbumDto;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.UUID;
+import java.util.stream.Collectors;
 
-public interface AlbumsService {
+@Service
+@RequiredArgsConstructor
+public class AlbumsService {
 
-    AlbumDto createAlbum(AlbumDto details) throws AlbumServiceException;
+    private final AlbumsRepository albumsRepository;
 
-    List<AlbumDto> findAlbumsByUserId(String userId);
+    private final AlbumMapper albumMapper;
 
-    List<AlbumDto> findAllAlbums();
+    public AlbumDto createAlbum(AlbumDto details) throws AlbumServiceException {
+        details.setAlbumId(UUID.randomUUID().toString());
 
-    AlbumDto findAlbumById(String id) throws AlbumNotFoundException;
+        Album album = albumMapper.map(details);
+        Album createdAlbum;
+        try {
+            createdAlbum = albumsRepository.save(album);
+        } catch (Exception e) {
+            throw new AlbumServiceException("Album creation failed: " + e.getMessage());
+        }
+
+        return albumMapper.map(createdAlbum);
+    }
+
+    public List<AlbumDto> findAlbumsByUserId(String userId) {
+        return albumsRepository.findByUserId(userId).stream()
+            .map(albumMapper::map)
+            .collect(Collectors.toList());
+    }
+
+    public List<AlbumDto> findAllAlbums() {
+        return albumsRepository.findAll().stream()
+            .map(albumMapper::map)
+            .collect(Collectors.toList());
+    }
+
+    public AlbumDto findAlbumById(String id) throws AlbumNotFoundException {
+        return albumsRepository.findById(id)
+            .map(albumMapper::map)
+            .orElseThrow(() -> new AlbumNotFoundException("Album not found"));
+    }
 }
